refacto: 🛠 Move some function from rundeck hook to helpers
