+++ lib/hooks/post/rundeck.js
@@ -1,35 +1,7 @@
 const https = require('https')
+const hookHelpers = require('../helpers')
 
 const RUNDECK_TOKEN = process.env.RUNDECK_TOKEN
-
-// Registry channels
-const DEV = 'dev'
-const BETA = 'beta'
-const STABLE = 'stable'
-
-const getChannel = version => {
-  if (version.includes('-dev.')) return DEV
-  if (version.includes('-beta.')) return BETA
-  if (version.match(/\d+\.\d+\.\d+/)) return STABLE
-  throw new Error('Unrecognized version channel')
-}
-
-const getInstanceDomain = instance => {
-  return instance
-    .split('.')
-    .slice(1)
-    .join('.')
-}
-
-const ENV_TARGETS = {
-  [DEV]: process.env.TARGETS_DEV || [],
-  [BETA]: (process.env.TARGETS_BETA || []).concat(
-    process.env.TARGETS_DEV || []
-  ),
-  [STABLE]: (process.env.TARGETS_STABLE || [])
-    .concat(process.env.TARGETS_BETA || [])
-    .concat(process.env.TARGETS_DEV || [])
-}
 
 const RUNDECK_JOBS = {
   'cozy.works': '87a668f7-eff1-422d-aa84-92b3bcc62c8f',
@@ -39,7 +11,7 @@ const RUNDECK_JOBS = {
 
 const runRundeckJob = (token, instance, slug) =>
   new Promise((resolve, reject) => {
-    const job = RUNDECK_JOBS[getInstanceDomain(instance)]
+    const job = RUNDECK_JOBS[hookHelpers.getInstanceDomain(instance)]
 
     if (!job) {
       console.log(`↳ ⚠️  No rundeck job available for ${instance}`)
@@ -81,22 +53,18 @@ module.exports = async options => {
   const { appSlug, appVersion } = options
 
   // Check if version is dev, beta, or stable.
-  const channel = getChannel(appVersion)
-  const targets = ENV_TARGETS[channel] && ENV_TARGETS[channel].replace(' ', '')
+  const channel = hookHelpers.getChannel(appVersion)
+  const instances = hookHelpers.getEnvInstances(channel)
 
-  if (!targets) {
+  if (!instances) {
     console.log(`↳ ℹ️  No instance to upgrade for channel ${channel}`)
     return options
   }
 
   console.log(
-    `↳ ℹ️  Updating channel ${channel} in instances ${targets.replace(
-      ',',
-      ', '
-    )}`
+    `↳ ℹ️  Updating channel ${channel} in instances ${instances.join(', ')}`
   )
 
-  const instances = targets.split(',')
   const failedDeployments = []
 
   for (const instance of instances) {
